refactor: move pure ui components from common to ui directory
