+++ frontend/src/App.tsx
@@ -9,7 +9,7 @@ import { SolanaAnchorWallet } from '@chains/solana/anchor-wallet';
 import { SolanaAuthSigner } from '@chains/solana/auth-signer';
 import { SolanaWalletProvider } from '@chains/solana/provider';
 import { solanaNetworkNameFromCluster } from '@constants/chains';
-import { ToastProvider } from '@components/common/toast';
+import { ToastProvider } from '@components/ui/toast';
 import { DynamicProvider } from '@contexts/dynamic';
 import { AppRoutes } from '@router';
 import { API_URL } from './config';
