+++ frontend/src/components/common/index.ts
@@ -1,5 +1,4 @@
 export { default as AuthActionButton } from './auth-action-button';
-export { default as NeonButton } from './neon-button';
-export { default as NeonCard } from './neon-card';
-export { default as NeonModal } from './neon-modal';
 export { default as TransactionStatus } from './transaction-status';
+// ui primitives re-exported for convenience
+export { NeonButton, NeonCard, NeonModal } from '@components/ui';
