+++ frontend/src/components/wallet/account-dropdown/index.tsx
@@ -5,8 +5,8 @@ import { useDynamicContext } from '@dynamic-labs/sdk-react-core';
 import { useAuth } from '@shared/core';
 import { getPopularTokens } from '@constants/tokens';
 import { Tones } from '@constants/tones';
-import { NeonButton, NeonCard } from '@components/common';
-import Icon, { CheckIcon, CopyIcon } from '@components/common/icon';
+import { NeonButton, NeonCard } from '@components/ui';
+import Icon, { CheckIcon, CopyIcon } from '@components/ui/icon';
 import { EthereumNetworkSwitcher, SolanaNetworkSwitcher } from '@components/wallet/network-switcher';
 import TokenBalance from '@components/wallet/token-balance';
 import NativeBalance from '@components/wallet/native-balance';
